refactor: Add validation decorators
- Validate user: user_logged_in.
- Validate post: post_exists, user_owns_post.
- Validate comment: comment_exists, user_owns_comment.

+++ handlers/like_handler.py
@@ -1,23 +1,20 @@
 import time
 
 from blog_handler import BlogHandler
-from models import Post, Like
+from models import Like
+from utility import user_logged_in, post_exists
 
 
 class LikeBtn(BlogHandler):
-    def post(self, post_id):
-        post = Post.by_id(int(post_id))
+    @user_logged_in
+    @post_exists
+    def post(self, post):
         like_btn = self.request.get('like-btn')
-        if post.user != self.user and post:
-            like = self.user.user_likes.filter("post =", post).get()
-            if like_btn == 'like':
-                if not like:
-                    like = Like.create(self.user, post)
-                    like.put()
-            elif like_btn == 'unlike':
-                if like:
-                    like.delete()
-            time.sleep(0.1)
-            self.redirect('/blog/' + post_id)
-        else:
-            self.error(403)
+        like = self.user.user_likes.filter("post =", post).get()
+        if like_btn == 'like' and not like:
+            like = Like.create(self.user, post)
+            like.put()
+        elif like_btn == 'unlike' and like:
+            like.delete()
+        time.sleep(0.1)
+        self.redirect('/blog/' + str(post.key().id()))
